pyproject.toml
[project]
name = "ylabcore"
version = "0.1.0"
description = "Laboratory instrument orchestration sandbox"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "paho-mqtt>=2.1.0",
    "pydantic>=2.12.4",
    "pyserial>=3.5",
    "pyyaml>=6.0.3",
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
]

[tool.setuptools.packages.find]
include = [
    "apps*",
    "core*",
]
exclude = [
    "tests*",
    "infra*",
    "configs*",
    "schemas*",
    "tools*",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
